+++ src/main/java/pl/agh/edu/erasmus_system/model/Edition.java
@@ -18,5 +18,5 @@ public class Edition {
     private Long id;
 
     @Column(nullable = false)
-    private Long year;
+    private Integer year;
 }
